feat(blog): 完善博客系统核心功能
后端：
- 完善文章管理模块（ArticleAppService、Repository）
- 新增分类命令对象（CreateCategoryCommand/UpdateCategoryCommand）
- 完善配置管理、技术日志、友链、项目、技能、标签等模块
- 强化Controller权限控制（Sa-Token）
- 完善DTO校验（JSR-303）
- 新增各类Command/Query对象规范入参

前端：
- 完善主题切换系统（ThemeSwitcher、useTheme）
- 优化文章卡片、日志卡片、项目卡片组件
- 完善管理后台各功能页面（文章/分类/标签/技能/项目/日志）
- 新增API类型定义和接口
- 优化样式系统（SCSS/Tailwind）

文档：
- 更新 CLAUDE.md 项目规范

+++ backend/src/main/java/com/nanmuli/blog/application/article/query/ArticlePageQuery.java
@@ -8,11 +8,11 @@ import lombok.Data;
 public class ArticlePageQuery {
 
     @Min(value = 1, message = "页码不能小于1")
-    private Integer current = 1;
+    private Long current = 1L;
 
     @Min(value = 1, message = "每页数量不能小于1")
     @Max(value = 100, message = "每页数量不能超过100")
-    private Integer size = 10;
+    private Long size = 10L;
 
     private Long categoryId;
     private Long tagId;
